requirements.txt
flask==2.3.3
openai==1.6.1
python-dotenv==1.0.0
pillow==10.0.0
pyheif==0.7.0